refactor: Add LottoMatchResultTest
